create detailTemp.js for testing
